Rename AsAIAgent to GetAIAgent and update for foundry as well
